Lepší zpracování finálních výsledků

+++ app/Services/GoogleCalendarService.php
@@ -43,9 +43,13 @@ class GoogleCalendarService
             $this->calendar->events->get('primary', $syncEvent->calendar_event_id); // Získání události z Google Kalendáře
             $this->calendar->events->delete('primary', $syncEvent->calendar_event_id); // Smazání události z Google Kalendáře
         } catch (\Exception $e) {
-            dd($e->getMessage());
+            //dd($e->getMessage());
         }
-        $syncEvent->delete();
+
+        if($syncEvent) {
+            $syncEvent->delete();
+        }
+
     }
 
     private function refreshToken($user)
